PZ-195,PZ-197 | Notification window should not be resizable; Decouple UI from license and provider implementation dependencies \PZ-197: + LicenseInfo, LicenseService model and interface contract respectively have been moved to common archive package + ZipState, FrmLicenseOverviewControllerm, ZipLauncher are decoupled from the license package implementation via the use of interfaces and archive package common objects. Implementation is derived via reflection. + Removed explicit dependencies on szjb,acc packages as service loader retrieves them.
PZ-195:
+ Notification stage is not resizeable.
