+++ pearl-zip-license/src/main/java/module-info.java
@@ -1,15 +1,14 @@
-import com.ntak.pearlzip.license.pub.LicenseService;
+import com.ntak.pearlzip.archive.pub.LicenseService;
 import com.ntak.pearlzip.license.pub.PearlZipLicenseService;
 
 /*
- * Copyright © 2021 92AK
+ * Copyright © 2022 92AK
  */
 /**
  *  Provides license information of core dependencies of the Pearl Zip project.
  */
 module com.ntak.pearlzip.license {
     exports com.ntak.pearlzip.license.pub;
-    exports com.ntak.pearlzip.license.model;
 
     provides LicenseService with PearlZipLicenseService;
