+++ pearl-zip-ui/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 #
-# Copyright © 2021 92AK
+# Copyright © 2022 92AK
 #
 
 configuration.ntak.pearl-zip.provider.priority.enabled.com.ntak.pearlzip.archive.szjb.pub.SevenZipArchiveService=true
@@ -17,6 +17,7 @@ configuration.ntak.pearl-zip.commit-hash=:: HASH ::
 configuration.ntak.pearl-zip.concurrency.lock-poll-timeout=100
 configuration.ntak.pearl-zip.default-format=zip
 configuration.ntak.pearl-zip.launcher-canonical-name=com.ntak.pearlzip.ui.mac.MacPearlZipApplication
+configuration.ntak.pearl-zip.license-service-canonical-name=com.ntak.pearlzip.license.pub.PearlZipLicenseService
 configuration.ntak.pearl-zip.jdbc.url=jdbc:oracle:thin:@(description= (retry_count=20)(retry_delay=3)(address=(protocol=tcps)(port=1522)(host=adb.uk-london-1.oraclecloud.com))(connect_data=(service_name=ga9b5a5185cc7a2_ntaknotify_low.adb.oraclecloud.com))(security=(ssl_server_cert_dn=\"CN=adwc.eucom-central-1.oraclecloud.com, OU=Oracle BMCS FRANKFURT, O=Oracle Corporation, L=Redwood City, ST=California, C=US\")))
 configuration.ntak.pearl-zip.jdbc.user=APP
 configuration.ntak.pearl-zip.jdbc.password=PZAUser0003Plus
